Update compare directories to output the diff.
