+++ steps/helm/test/integration/01-with-multiple-services-and-children/01-with-multiple-services-and-children.spec.ts
@@ -1,42 +1,13 @@
-import FileSystemSync from 'node:fs';
-import FileSystem from 'node:fs/promises';
 import Path from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { BaumManager } from '@veto-party/baum__core';
 import { NPMPackageManager } from '@veto-party/baum__package_manager__npm';
 import { HelmGenerator } from '../../../src/HelmGenerator.js';
 import { HelmGeneratorProvider } from '../../../src/HelmGeneratorProvider.js';
+import { compareDirectories } from '../utility/compareDirectories.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = Path.dirname(__filename);
-
-async function compareDirectories(pathA: string, pathB: string) {
-  const filesA = FileSystemSync.readdirSync(pathA);
-  const filesB = FileSystemSync.readdirSync(pathB);
-
-  if (filesA.length !== filesB.length) {
-    console.error(`File count is different....  ${JSON.stringify(filesA)} ${JSON.stringify(filesB)}`);
-    return false;
-  }
-
-  for (const file of filesA) {
-    if ((await FileSystem.stat(Path.join(pathA, file))).isDirectory()) {
-      if (!(await compareDirectories(Path.join(pathA, file), Path.join(pathB, file)))) {
-        return false;
-      }
-    } else {
-      const fileA = FileSystemSync.readFileSync(Path.join(pathA, file), 'utf-8');
-      const fileB = FileSystemSync.readFileSync(Path.join(pathB, file), 'utf-8');
-
-      if (fileA !== fileB) {
-        console.error(`File ${file} is different`);
-        return false;
-      }
-    }
-  }
-
-  return true;
-}
 
 describe('A 01-with-multiple-services-and-children', () => {
   it('Should run successfully', async () => {
